requirements.txt
pandas==2.2.2
requests==2.32.3
Pillow==10.4.0
openpyxl==3.1.2
